% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_closest_station.R
\name{closest_station}
\alias{closest_station}
\title{Compute the closet station given current location}
\usage{
closest_station(long, lat)
}
\arguments{
\item{long}{longtitude of user location}

\item{lat}{latitude of user location}
}
\value{
a tbl_df object showing the distance between the user and top five closest stations with ID, name, number of docks, and position
}
\description{
This function returns stations within walking distance given the user's current location
}
\examples{
# Calculate distance for user at (-71.11467361, 42.34414899)
top_5_station <- closest_station(-71.11467361, 42.34414899)
}
