% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_data.R
\docType{data}
\name{station_data}
\alias{station_data}
\title{Blue bike station data}
\format{
A data frame of 423 rows and 8 columns
\describe{
\item{number}{Station ID}
\item{name}{Station name}
\item{latitude}{Latitude of the station}
\item{longitude}{Longitude of the station}
\item{district}{District of the station}
\item{public}{Character vector showing if a station is public}
\item{total_docks}{The number of docks at each station}
\item{deployment_year}{The year that the station was put into work}
}
}
\source{
The original source of the data are bluebikes system data retrieved from \url{https://www.bluebikes.com/system-data}
}
\usage{
station_data
}
\description{
A dataset that includes identification, position, and other basic information about bluebike stations
}
\keyword{datasets}
