% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_distance.R
\name{station_distance}
\alias{station_distance}
\title{Compute the distance from stations given current location}
\usage{
station_distance(long, lat)
}
\arguments{
\item{long}{longtitude of user location}

\item{lat}{latitude of user location}
}
\value{
a tbl_df object showing the distance between the user and top five closest stations with ID, name, number of docks, and position
}
\description{
This function returns stations with distance in ascending order given the user's current location
}
\examples{
# Calculate distance for user at (-71.11467361, 42.34414899)
top_5_station <- station_distance(-71.11467361, 42.34414899)
}
