% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_radius.R
\name{station_radius}
\alias{station_radius}
\title{Plot bike stations within a given radius}
\usage{
station_radius(long, lat, r = 1000)
}
\arguments{
\item{long}{numeric value of longitude}

\item{lat}{numeric value of latitude}

\item{r}{numeric value of set radius in meters}
}
\value{
A leaflet map
}
\description{
This function plots the position of the stations within walking distance
}
\examples{
# Show user at (-71.11467, 42.34415) and set the radius to 500 m
station_radius(long = -71.11467, lat = 42.34415, r = 2000)
}
