% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_distance.R
\name{trip_distance}
\alias{trip_distance}
\title{Compute trip distance for a specific dataset}
\usage{
trip_distance(data)
}
\arguments{
\item{data}{trip data pulled from the Blue Bike System data}
}
\value{
a tbl_df object with an additional distance column
}
\description{
This function computes the geographical distance between the start and end stations for trips in a given dataset
}
\examples{
# Calculate distance for sample trip data
sample_distance <- trip_distance(trip_history_sample)$distance
}
