% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bluebike.R
\docType{package}
\name{bluebike}
\alias{bluebike}
\title{bluebike - A Data Package for Bluebike users}
\description{
The bluebike package provides the sampled dataset from Bluebike System Data
}
\examples{
library(bluebike)
library(tidyverse)

# Find most used stations:
stations <- trip_history_sample \%>\%
    group_by(`start station name`) \%>\%
    summarize(trips_from = n())
head(stations)

}
